Refactored App code- split Header into its own component to ensure easier updates of main page
