+++ src/components/App/App.css
@@ -1,71 +1,26 @@
-.App {
-  display: flex;
-  flex-direction: column; 
-  background-color: #37312A;
-  min-height: 100vh;
-  margin: 0;
-  padding: 0;
-}
-
 body {
   background-color: #37312A;
+  margin: 0;
+  height: 100%;
 }
 
-header {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  position: fixed;
-  width: 100%;
-  top: 0;
-  z-index: 1000;
-  padding: 10px 0;
-}
-
-header img {
-  margin-left: 10px;
-  margin-top: 10px;
-  margin-right: 10px;
-  height: 100px;
-  border-radius: 30px;
-}
-
-.burger-img {
-  height: 50px;
-}
-
-header h1 {
-  margin: auto;
-}
-
-.burger-button:hover {
- cursor: pointer
-}
-
-.pop-up {
+.App {
   display: flex;
   flex-direction: column;
-  border-bottom-right-radius: 10px;
-  border-bottom-left-radius: 10px;
-  background-color: transparent;
-  color: #00C9AF;
-  text-align: justify;
-  padding: 10px;
-}
-
-.pop-out-container {
-  margin: 10px;
-  padding: 10px;
-  display: flex;
-  flex-direction: column;
+  background-color: #37312A;
+  min-height: 100vh; 
+  justify-content: center; 
 }
 
 .main-content {
   display: flex;
   flex-direction: column;
+  justify-content: center; 
+  align-items: center;
   text-align: center;
+  flex-grow: 1; 
   width: 100%;
-  margin-top: 250px; 
+  padding: 20px;
 }
 
 .searchbar-container,
