+++ src/components/Header/Header.css
@@ -1,0 +1,57 @@
+header {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  position: fixed;
+  width: 100%;
+  top: 0;
+  z-index: 1000;
+  padding: 10px 0;
+}
+
+header img {
+  margin-left: 10px;
+  margin-top: 10px;
+  margin-right: 10px;
+  height: 100px;
+  border-radius: 30px;
+}
+
+header h1 {
+  margin: auto;
+}
+
+.burger-img {
+  height: 50px;
+}
+
+.burger-button:hover {
+ cursor: pointer
+}
+
+.pop-up {
+  display: flex;
+  flex-direction: column;
+  border-bottom-right-radius: 10px;
+  border-bottom-left-radius: 10px;
+  background-color: transparent;
+  color: #00C9AF;
+  text-align: justify;
+  padding: 10px;
+}
+
+.pop-out-container {
+  margin: 10px;
+  padding: 10px;
+  display: flex;
+  flex-direction: column;
+}
+
+.pop-up ul {
+  list-style-type: none;
+  padding: 0;
+}
+
+.pop-up li {
+  margin-bottom: 10px;
+}
